requirements.txt
langchain
gradio
python-dotenv
nltk 
chromadb
langchain-groq
huggingface_hub
langchain-community
pypdf
